Fix: GET /product route not responding due to incorrect service injection
This commit resolves the issue where the GET /product route was returning a 404 error with "Requested application data is not configured correctly."

Key changes include:
- Refactored service injection: Modified  in  and  to correctly initialize and provide the  instance using , passing  for database pool access. This maintains modularity while ensuring proper dependency injection.
- Corrected : Replaced  with  in  to use the correct  method for length checking.
- Updated : Adjusted the call to  to pass the  data.

These changes ensure the  is properly available to handlers, allowing the /product route to function as expected.

+++ src/app/module/product/models.rs
@@ -55,7 +55,7 @@ pub(crate) mod model {
             #[validate(length(min = 3,max = 50))]
             pub slug: String,
             #[validate(range(min = 0.0))]
-            pub price: f32,
+            pub price: f64,
             #[validate(range(min = 0))]
             pub stock: i32,
             pub images: Vec<String>,
